pyproject.toml
[project]
name = "portfolio-mgmt-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
